2023_08_08

+++ src/main/java/com/choongang/OriMarket/user/UserService.java
@@ -4,9 +4,11 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.ui.Model;
+import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpSession;
-import java.util.Optional;
+import java.util.*;
 
 @Service
 @RequiredArgsConstructor 
@@ -15,12 +17,13 @@ public class UserService {
 
     @Autowired
     private final UserRepository userRepository;
+    private final UserAddressRepository userAddressRepository;
 
-    public boolean login(User member,HttpSession session) {
-
+    public boolean login(User member, HttpSession session,Model model) {
+        session.removeAttribute("userAddress1");
+        session.removeAttribute("userAddressDetail1");
         //값이 null일 떄 Optinal이 처리
         User findUser = userRepository.findByUserId(member.getUserId());
-
 
         if(findUser == null){
             return false;
@@ -29,6 +32,17 @@ public class UserService {
         if(!findUser.getUserPassword().equals(member.getUserPassword())){
             return false;
         }
+        List<UserAddress> userAddresses = findUser.getUserAddresses();
+        //ModelAndView mav = new ModelAndView();
+        if (userAddresses.isEmpty()){
+        }else {
+            //mav.addObject("userAd",userAddresses);
+            //mav.setViewName("main/search");
+            session.setAttribute("userAddress1",userAddresses.get(userAddresses.size()-1).getUserAddress1());
+            session.setAttribute("userAddressDetail1",userAddresses.get(userAddresses.size()-1).getUserAddressDetail1());
+        }
+
+
         session.setAttribute("userSeq",findUser.getUserSeq());
         session.setAttribute("userName",findUser.getUserName());
         session.setAttribute("userId",findUser.getUserId());
@@ -67,10 +81,23 @@ public class UserService {
        //userRepository.findByUserId()
    }
 
-   public void delete(Long UserSeq){
-        User user = userRepository.findById(UserSeq).orElseThrow();
-        userRepository.delete(user);
-   }
+    public boolean delete(Long UserSeq) {
+        User user = userRepository.findById(UserSeq).orElse(null);
+        if (user != null) {
+            List<UserAddress> userAddresses = user.getUserAddresses();
+            try {
+                for (UserAddress userAddress : userAddresses) {
+                    userAddressRepository.delete(userAddress);
+                }
+                userRepository.delete(user);
+                return true;
+            } catch (Exception e) {
+
+                e.printStackTrace();
+            }
+        }
+        return false;
+    }
 
    public boolean checkUserId(String userId){
        return userRepository.existsByUserId(userId);
